requirements.txt
aiohttp==3.8.2
python-dotenv
alpaca-py
fastapi
uvicorn
jinja2
typer[all]
python-multipart